Вносит правки, переносит фильтрацию в утилиты, правит шаблон
