+++ controllers/TasksController.php
@@ -4,33 +4,33 @@ namespace app\controllers;
 
 use Yii;
 use yii\web\Controller;
-use app\models\SearchForm;
-use app\services\TaskService;
-use app\services\CategoryService;
+use app\models\TasksSearchForm;
+use app\models\Tasks;
+use app\models\Categories;
+use TaskForce\utils\TasksFilter;
 
 class TasksController extends Controller
 {
     public function actionIndex()
     {
-        $model = new SearchForm();
+        $model = new TasksSearchForm();
 
         if (Yii::$app->request->getIsPost()) {
             $model->load(Yii::$app->request->post());
 
             if ($model->validate()) {
-                $tasks = (new TaskService())->getFilteredTasks($model);
+                $tasks = (new TasksFilter())->getFilteredTasks($model);
             }
         }
 
-        !isset($tasks) && $tasks = (new TaskService())->getAllTasks();
-
-        $categories = (new CategoryService())->getAllCategories();
+        !isset($tasks) && $tasks = Tasks::find()->all();
+        $categories = Categories::find()->all();
 
         return $this->render('index', [
             'model' => $model,
             'tasks' => $tasks,
             'categories' => $categories,
-            'period_values' => SearchForm::PERIOD_VALUES
+            'period_values' => TasksSearchForm::PERIOD_VALUES
         ]);
     }
 }
